worker/requirements.txt
numpy==1.24.2
opencv-python==4.7.0.72
pdf2image==1.16.2
Pillow==9.4.0
pika==1.3.1
pymongo==4.3.3
oci==2.94.0
python-dotenv=1.0.0